% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{colors}
\alias{colors}
\alias{colors_license}
\alias{colors_license_unpaywall}
\alias{colors_source_disclosure}
\title{Color palettes
Colors used in the hybrid open access dashboard, based on \code{\link[metar:ugoe_pal]{metar::ugoe_pal()}}}
\format{
An object of class \code{character} of length 6.

An object of class \code{character} of length 2.

An object of class \code{character} of length 4.
}
\usage{
colors_license

colors_license_unpaywall

colors_source_disclosure
}
\description{
Color palettes
Colors used in the hybrid open access dashboard, based on \code{\link[metar:ugoe_pal]{metar::ugoe_pal()}}
}
\section{Functions}{
\itemize{
\item \code{colors_license}: licenses

\item \code{colors_license_unpaywall}: other types of OA license information detected by Unpaywall

\item \code{colors_source_disclosure}: sources of disclosure of OA sponsorship
}}

\concept{visualisation functions}
\keyword{datasets}
