% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_all_ints}
\alias{find_all_ints}
\title{Find all integers inside the range of a vector}
\usage{
find_all_ints(x)
}
\arguments{
\item{x}{a numeric vector that can be transformed to integers (integerish)}
}
\value{
an integer vector from min to max
}
\description{
Helpful for finding all years.
\code{\link[=unique]{unique()}} is't quite the same, because it would leave out empty years, but for plots and the like there can never be empty years.
}
