% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoad.R
\docType{package}
\name{hoad-package}
\alias{hoad}
\alias{hoad-package}
\title{hoad: Hybrid Open Access Dashboard}
\description{
Analytics to increase cost transparency in business model transformation towards open access publishing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://subugoe.github.io/hoad}
  \item \url{https://github.com/subugoe/hoad/}
  \item Report bugs at \url{https://github.com/subugoe/hoad/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Authors:
\itemize{
  \item Najko Jahn \email{najko.jahn@gmail.com} (\href{https://orcid.org/0000-0001-5105-1463}{ORCID})
  \item Jan Höffler \email{hoeffler@sub.uni-goettingen.de} (\href{https://orcid.org/0000-0002-8382-2071}{ORCID})
  \item Henrieke Walter \email{henrieke.walter@sub.uni-goettingen.de} (\href{https://orcid.org/0000-0003-3313-7364}{ORCID})
}

Other contributors:
\itemize{
  \item Göttingen State and University Library [copyright holder, funder]
  \item German Research Foundation [sponsor]
  \item Crossref [data contributor]
  \item OpenAPC Initiative [data contributor]
}

}
\keyword{internal}
