% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hybrid_publications}
\alias{hybrid_publications}
\title{Hybrid open access articles via \href{https://www.crossref.org}{Crossref}}
\format{
A data frame with the following variables:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{license} \tab Normalized open content license statement \cr
   \code{journal_title} \tab Most frequent journal title used by Crossref \cr
   \code{publisher} \tab Most frequent publisher name used by Crossref \cr
   \code{doi_oa} \tab DOI of the hybrid open access article \cr
   \code{issued} \tab Earliest publication year \cr
   \code{yearly_jn_volume} \tab Yearly article volume per journal \cr
   \code{license_ref_n} \tab Yearly article volume under the license \code{license} \cr
   \code{yearly_publisher_volume} \tab Yearly article volume of all journals in the dataset per publisher \cr
   \code{yearly_all} \tab Yearly article volume investigated \cr
   \code{period} \tab Year of reporting to Open APC \cr
   \code{euro} \tab The amount that was paid in EURO. Includes VAT and additional fees \cr
   \code{hybrid_type} \tab Spending source (Open APC, SCOAP3) \cr
   \code{institution} \tab Top-level organisation which reported article to Open APC \cr
   \code{country} \tab Country of origin (iso3c) \cr
   \code{country_name} \tab Country of origin (name) \cr
   \code{license_url} \tab License URL \cr
   \code{host} \tab Email host first or corresponding author \cr
   \code{subdomain} \tab Email subdomain first or corresponding author \cr
   \code{domain} \tab Email domain first or corresponding author \cr
   \code{suffix} \tab Email suffix first or corresponding author \cr
}
}
\source{
\href{https://www.crossref.org}{Crossref}
}
\usage{
hybrid_publications()
}
\description{
Contains information about the overall publication volume, and, if available, cost information from the Open APC Initiative.
}
\section{License}{

See Crossref \href{https://www.crossref.org/requestaccount/termsandconditions.html}{Terms and Conditions}
}

\seealso{
Other data: 
\code{\link{unpaywall_df}()}
}
\concept{data}
