% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{oa_by_license}
\alias{oa_by_license}
\alias{oa_by_license_plot}
\title{Open Access Articles by License and Year}
\usage{
oa_by_license()

oa_by_license_plot()
}
\description{
Open Access Articles by License and Year
}
\section{Functions}{
\itemize{
\item \code{oa_by_license_plot}: Plot by License and Year
}}

\examples{
oa_by_license()
oa_by_license_plot()
}
\concept{ETL functions}
\concept{visualisation function_}
