% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{oa_by_other}
\alias{oa_by_other}
\alias{oa_by_other_plot}
\title{Open Access Articles by Year and License Statement}
\usage{
oa_by_other(x = unpaywall_df(), by_pub = FALSE)

oa_by_other_plot(x = oa_by_other())
}
\arguments{
\item{x}{a data frame like its default, \code{\link[=oa_by_other]{oa_by_other()}}}

\item{by_pub}{giving whether to group by summarised publishers}
}
\description{
\emph{Additional} license information (compared to \code{\link[=oa_by_license]{oa_by_license()}}) as detected by Unpaywall.
}
\section{Functions}{
\itemize{
\item \code{oa_by_other_plot}: plotting
}}

\examples{
oa_by_other()
oa_by_other_plot()
}
\seealso{
Other ETL functions: 
\code{\link{oa_by_license}()}
}
\concept{ETL functions}
\concept{visualisation function}
