% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{path_extdat}
\alias{path_extdat}
\title{Find path for files in \verb{/inst/extdata/}}
\usage{
path_extdat(x)
}
\arguments{
\item{x}{\verb{[character(1)]} giving the filename or path to the file from \verb{inst/extdata/}}
}
\value{
\verb{[character(1)]} giving the absolute path to \code{x}, as returned by \code{\link[fs:path_math]{fs::path_abs()}}.
}
\description{
Looks for \verb{inst/} in the working directory, otherwise wraps \code{\link[fs:path_package]{fs::path_package()}} with appropriate defaults for this package.
}
\details{
This function \strong{will be deprecated} when all data files are retrieved via APIs.

As a \url{stop-gap measure}, data files are stored in \verb{inst/extdata/} and compiled into the package.
Depending on the context in which you want to use these files, their paths will differ.
When running from the source root, they can be found at \verb{inst/extdata/}.
When the package is compiled (as in vignette use), they can be found at \verb{extadata/}.

This function automatically picks the correct path or throws an error if the file cannot be found.
}
\keyword{internal}
