% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-data.R
\docType{data}
\name{unpaywall_df}
\alias{unpaywall_df}
\title{Unpaywall data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 43892 rows and 5 columns.
}
\usage{
unpaywall_df
}
\description{
Unpaywall data
}
\keyword{datasets}
