% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{oa_by_license}
\alias{oa_by_license}
\alias{oa_by_license_plot}
\title{Open Access Articles by Year and License}
\usage{
oa_by_license()

oa_by_license_plot()
}
\description{
Open Access Articles by Year and License
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{oa_by_license_plot}: Plot by License and Year
}}}

\examples{
oa_by_license()
oa_by_license_plot()
}
\seealso{
Other ETL functions: 
\code{\link{oa_by_other}()}
}
\concept{ETL functions}
\concept{visualisation function_}
